pipeline/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
anthropic>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.1
